api_project/requirements.txt
Django>=5.0.0,<6.0.0
djangorestframework>=3.15.0,<4.0.0
